requirements.txt
numpy~=2.2.5
azure-iot-device~=2.14.0
tqdm~=4.67.1
python-dotenv~=1.1.0
awsiot~=0.1.3
awscrt~= 0.27.4
requests~=2.32.3